app/requirements.txt
flask-pymongo==2.3.0
flask-unittest==0.1.3
mongomock==4.1.2
types-waitress==2.1.4.4
waitress==2.1.2
